+++ test/HesperidesIntegrationSpec.groovy
@@ -414,7 +414,7 @@ class HesperidesIntegrationSpec extends Specification implements Helper {
     def "Can create a new module in workingcopy from a module in workingcopy"() {
         when:
             hesperides.createModule(moduleName: 'toto', version: '0.0')
-            hesperides.createModule(moduleName: 'toto', version: '0.1', fromModule: [name: 'toto', version: '0.0', isWorkingcopy: true])
+            hesperides.createModule(moduleName: 'toto', version: '0.1', fromModule: [name: 'toto', version: '0.0', isWorkingCopy: true])
 
         then:
             hesperides.getModule(moduleName: 'toto', version: '0.1', moduleType: 'workingcopy')
@@ -422,6 +422,19 @@ class HesperidesIntegrationSpec extends Specification implements Helper {
         cleanup:
             hesperides.deleteModule(moduleName: 'toto', version: '0.1', moduleType: 'workingcopy')
             hesperides.deleteModule(moduleName: 'toto', version: '0.0', moduleType: 'workingcopy')
+    }
+
+    def "Can create a new module in workingcopy from a module in workingcopy (deprecated isWorkingcopy argument)"() {
+        when:
+        hesperides.createModule(moduleName: 'toto', version: '0.0')
+        hesperides.createModule(moduleName: 'toto', version: '0.1', fromModule: [name: 'toto', version: '0.0', isWorkingcopy: true])
+
+        then:
+        hesperides.getModule(moduleName: 'toto', version: '0.1', moduleType: 'workingcopy')
+
+        cleanup:
+        hesperides.deleteModule(moduleName: 'toto', version: '0.1', moduleType: 'workingcopy')
+        hesperides.deleteModule(moduleName: 'toto', version: '0.0', moduleType: 'workingcopy')
     }
 
     def "Can create a new module in workingcopy from a released module"() {